Eliminate a compiler warning on OmniOS.
* test/src/unix/tester-check-net.cc: pid_t is a long int on OmniOS, so
  consistently convert to long int and use %ld for fprintf.

+++ test/src/unix/tester-check-net.cc
@@ -253,7 +253,8 @@ int main(void)
     }
   if (p != child)
     {
-      fprintf(stderr, "teardown: unexpected child %d != %d\n", p, child);
+      fprintf(stderr, "teardown: unexpected child %ld != %ld\n",
+              (long int) p, (long int) child);
       return 2;
     }
   if (!WIFEXITED(status))
